// src/common/enums/language.enum.ts
export enum LanguageEnum {
  PT_BR = "PT_BR", // Português do Brasil
  EN_US = "EN_US", // Inglês (Estados Unidos)
  ES_ES = "ES_ES", // Espanhol (Espanha)
  FR_FR = "FR_FR", // Francês (França)
  DE_DE = "DE_DE", // Alemão (Alemanha)
  ZH_CN = "ZH_CN", // Mandarim (China)
  JA_JP = "JA_JP", // Japonês (Japão)
  KO_KR = "KO_KR", // Coreano (Coreia do Sul)
}